Migrate formatting_html.py into xarray core (#8930)
* DAS-2066: formatting_html.py merge

* Revert "DAS-2066: formatting_html.py merge"
oopsies
This reverts commit d68ae3c823bbbe26c918605a388464b5ba2309c2.

* Adding datatree functions to core

* Deleted formatting_html.py and test_formatting_html.py in datatree_

* renamed all of the relevant functions so they represent datatree

* added comment for mypy

* added dict[any, any] as type hint for children

* added dict[Any, Any] as type hint for children

* updated typhint to dict[str, DataTree] and increased pixel min width to 700

* replaced Any with DataTree

* added type checking

* DAS-2066 - Incorporate CSS changes for DataTree display.

* DAS-2066 - Fix tests.

---------
