+++ xarray/core/formatting_html.py
@@ -2,9 +2,11 @@ from __future__ import annotations
 
 import uuid
 from collections import OrderedDict
+from collections.abc import Mapping
 from functools import lru_cache, partial
 from html import escape
 from importlib.resources import files
+from typing import TYPE_CHECKING
 
 from xarray.core.formatting import (
     inline_index_repr,
@@ -17,6 +19,9 @@ STATIC_FILES = (
     ("xarray.static.html", "icons-svg-inline.html"),
     ("xarray.static.css", "style.css"),
 )
+
+if TYPE_CHECKING:
+    from xarray.core.datatree import DataTree
 
 
 @lru_cache(None)
@@ -341,3 +346,129 @@ def dataset_repr(ds) -> str:
     ]
 
     return _obj_repr(ds, header_components, sections)
+
+
+def summarize_datatree_children(children: Mapping[str, DataTree]) -> str:
+    N_CHILDREN = len(children) - 1
+
+    # Get result from datatree_node_repr and wrap it
+    lines_callback = lambda n, c, end: _wrap_datatree_repr(
+        datatree_node_repr(n, c), end=end
+    )
+
+    children_html = "".join(
+        (
+            lines_callback(n, c, end=False)  # Long lines
+            if i < N_CHILDREN
+            else lines_callback(n, c, end=True)
+        )  # Short lines
+        for i, (n, c) in enumerate(children.items())
+    )
+
+    return "".join(
+        [
+            "<div style='display: inline-grid; grid-template-columns: 100%; grid-column: 1 / -1'>",
+            children_html,
+            "</div>",
+        ]
+    )
+
+
+children_section = partial(
+    _mapping_section,
+    name="Groups",
+    details_func=summarize_datatree_children,
+    max_items_collapse=1,
+    expand_option_name="display_expand_groups",
+)
+
+
+def datatree_node_repr(group_title: str, dt: DataTree) -> str:
+    header_components = [f"<div class='xr-obj-type'>{escape(group_title)}</div>"]
+
+    ds = dt.ds
+
+    sections = [
+        children_section(dt.children),
+        dim_section(ds),
+        coord_section(ds.coords),
+        datavar_section(ds.data_vars),
+        attr_section(ds.attrs),
+    ]
+
+    return _obj_repr(ds, header_components, sections)
+
+
+def _wrap_datatree_repr(r: str, end: bool = False) -> str:
+    """
+    Wrap HTML representation with a tee to the left of it.
+
+    Enclosing HTML tag is a <div> with :code:`display: inline-grid` style.
+
+    Turns:
+    [    title    ]
+    |   details   |
+    |_____________|
+
+    into (A):
+    |─ [    title    ]
+    |  |   details   |
+    |  |_____________|
+
+    or (B):
+    └─ [    title    ]
+       |   details   |
+       |_____________|
+
+    Parameters
+    ----------
+    r: str
+        HTML representation to wrap.
+    end: bool
+        Specify if the line on the left should continue or end.
+
+        Default is True.
+
+    Returns
+    -------
+    str
+        Wrapped HTML representation.
+
+        Tee color is set to the variable :code:`--xr-border-color`.
+    """
+    # height of line
+    end = bool(end)
+    height = "100%" if end is False else "1.2em"
+    return "".join(
+        [
+            "<div style='display: inline-grid; grid-template-columns: 0px 20px auto; width: 100%;'>",
+            "<div style='",
+            "grid-column-start: 1;",
+            "border-right: 0.2em solid;",
+            "border-color: var(--xr-border-color);",
+            f"height: {height};",
+            "width: 0px;",
+            "'>",
+            "</div>",
+            "<div style='",
+            "grid-column-start: 2;",
+            "grid-row-start: 1;",
+            "height: 1em;",
+            "width: 20px;",
+            "border-bottom: 0.2em solid;",
+            "border-color: var(--xr-border-color);",
+            "'>",
+            "</div>",
+            "<div style='",
+            "grid-column-start: 3;",
+            "'>",
+            r,
+            "</div>",
+            "</div>",
+        ]
+    )
+
+
+def datatree_repr(dt: DataTree) -> str:
+    obj_type = f"datatree.{type(dt).__name__}"
+    return datatree_node_repr(obj_type, dt)
